refactor: rename generated api-client location
