+++ composables/use-geo-marker-layers.ts
@@ -1,7 +1,7 @@
 import { useQueries } from "@tanstack/vue-query";
 import type { Feature, Point } from "geojson";
 
-import type { MarkerProperties, MarkersType, QueryDescription } from "@/lib/api-client/Api";
+import type { MarkerProperties, MarkersType, QueryDescription } from "@/lib/api-client";
 
 export function useGeoMarkerLayers(
 	queries: MaybeRef<Array<Required<QueryDescription>>>,
