+++ package.json
@@ -15,7 +15,7 @@
 		"dev": "nuxt dev --dotenv ./.env.local",
 		"format:check": "prettier . --cache --check --ignore-path ./.gitignore",
 		"format:fix": "pnpm run format:check --write",
-		"generate:api-client": "sta --path \"https://raw.githubusercontent.com/acdh-oeaw/vicav-app-api/master/openapi.json\" --clean-output --output ./lib/api-client/",
+		"generate:api-client": "swagger-typescript-api --path \"https://raw.githubusercontent.com/acdh-oeaw/vicav-app-api/master/openapi.json\" --clean-output --name index.ts --output ./lib/api-client/",
 		"lint:check": "run-p --continue-on-error \"lint:*:check\"",
 		"lint:fix": "run-p --continue-on-error \"lint:*:fix\"",
 		"lint:code:check": "eslint . --cache --ext .js,.ts,.vue --ignore-path ./.gitignore",
@@ -40,14 +40,14 @@
 		"@acdh-oeaw/lib": "^0.1.6",
 		"@fontsource-variable/roboto-flex": "^5.0.8",
 		"@pinia/nuxt": "^0.4.11",
-		"@tanstack/vue-query": "^5.0.0-rc.6",
+		"@tanstack/vue-query": "^5.0.0-rc.11",
 		"@vueuse/core": "^10.5.0",
 		"@vueuse/nuxt": "^10.5.0",
 		"colorjs.io": "^0.4.5",
 		"cva": "^1.0.0-beta.1",
 		"is-ci": "^3.0.1",
 		"leaflet": "^1.9.4",
-		"lucide-vue-next": "^0.284.0",
+		"lucide-vue-next": "^0.286.0",
 		"npm-run-all2": "^6.1.1",
 		"nuxt": "^3.7.4",
 		"pinia": "^2.1.6",
@@ -63,9 +63,9 @@
 		"@acdh-oeaw/prettier-config": "^1.0.1",
 		"@acdh-oeaw/stylelint-config": "^1.0.6",
 		"@acdh-oeaw/tsconfig": "^1.0.1",
-		"@axe-core/playwright": "^4.7.3",
+		"@axe-core/playwright": "^4.8.0",
 		"@nuxt/devtools": "^1.0.0-beta.1",
-		"@playwright/test": "^1.38.1",
+		"@playwright/test": "^1.39.0",
 		"@tailwindcss/typography": "^0.5.10",
 		"@types/geojson": "^7946.0.11",
 		"@types/leaflet": "^1.9.6",
